refactor: keep shared MLS utilities independent of client storage
