+++ crates/xmtp_db/src/encrypted_store/group/convert.rs
@@ -1,8 +1,6 @@
 use super::*;
 use xmtp_proto::ConversionError;
-use xmtp_proto::xmtp::device_sync::group_backup::{ConversationTypeSave, GroupMembershipStateSave};
-
-use xmtp_proto::xmtp::mls::message_contents::ConversationType as ConversationTypeProto;
+use xmtp_proto::xmtp::device_sync::group_backup::GroupMembershipStateSave;
 
 impl TryFrom<GroupMembershipStateSave> for GroupMembershipState {
     type Error = ConversionError;
@@ -21,21 +19,6 @@ impl TryFrom<GroupMembershipStateSave> for GroupMembershipState {
     }
 }
 
-impl TryFrom<ConversationTypeSave> for ConversationType {
-    type Error = ConversionError;
-    fn try_from(value: ConversationTypeSave) -> Result<Self, Self::Error> {
-        let conversation_type = match value {
-            ConversationTypeSave::Dm => Self::Dm,
-            ConversationTypeSave::Group => Self::Group,
-            ConversationTypeSave::Sync => Self::Sync,
-            ConversationTypeSave::Unspecified => {
-                return Err(ConversionError::Unspecified("conversation_type"));
-            }
-        };
-        Ok(conversation_type)
-    }
-}
-
 impl From<GroupMembershipState> for GroupMembershipStateSave {
     fn from(value: GroupMembershipState) -> Self {
         match value {
@@ -45,53 +28,5 @@ impl From<GroupMembershipState> for GroupMembershipStateSave {
             GroupMembershipState::Restored => Self::Restored,
             GroupMembershipState::PendingRemove => Self::PendingRemove,
         }
-    }
-}
-impl From<ConversationType> for ConversationTypeSave {
-    fn from(value: ConversationType) -> Self {
-        match value {
-            ConversationType::Dm => Self::Dm,
-            ConversationType::Group => Self::Group,
-            ConversationType::Sync => Self::Sync,
-            ConversationType::Oneshot => Self::Unspecified,
-        }
-    }
-}
-
-/**
- * XMTP supports the following types of conversation
- *
- * *Group*: A conversation with 1->N members and complex permissions and roles
- * *DM*: A conversation between 2 members with simplified permissions
- * *Sync*: A conversation between all the devices of a single member with simplified permissions
- */
-impl From<ConversationType> for ConversationTypeProto {
-    fn from(value: ConversationType) -> Self {
-        match value {
-            ConversationType::Group => Self::Group,
-            ConversationType::Dm => Self::Dm,
-            ConversationType::Sync => Self::Sync,
-            ConversationType::Oneshot => Self::Oneshot,
-        }
-    }
-}
-
-impl TryFrom<i32> for ConversationType {
-    type Error = xmtp_proto::ConversionError;
-
-    fn try_from(value: i32) -> Result<Self, Self::Error> {
-        Ok(match value {
-            1 => Self::Group,
-            2 => Self::Dm,
-            3 => Self::Sync,
-            4 => Self::Oneshot,
-            n => {
-                return Err(ConversionError::InvalidValue {
-                    item: "ConversationType",
-                    expected: "number between 1 - 4",
-                    got: n.to_string(),
-                });
-            }
-        })
     }
 }
