+++ crates/xmtp_mls_common/Cargo.toml
@@ -22,11 +22,9 @@ prost.workspace = true
 serde.workspace = true
 thiserror.workspace = true
 tls_codec.workspace = true
-xmtp-workspace-hack.workspace = true
 xmtp_common.workspace = true
 xmtp_configuration.workspace = true
 xmtp_cryptography.workspace = true
-xmtp_db.workspace = true
 xmtp_id.workspace = true
 xmtp_proto.workspace = true
