+++ crates/xmtp_mls_common/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod app_data;
+pub mod commit_log;
 pub mod group;
 pub mod group_metadata;
 pub mod group_mutable_metadata;
